// types/projectConfig.ts
export interface ProjectConfig {
    type: string;
    language: string;
    bootVersion: string;
    groupId: string;
    artifactId: string;
    name: string;
    description: string;
    packageName: string;
    packaging: string;
    javaVersion: string;
    configurationFileFormat: string;
    baseDir: string;
    dependencies: string[];
    backlink: string; // computed from window.location
  }
  